requirements.txt
streamlit
PyPDF2
langchain
langchain-google-genai
google-generativeai
faiss-cpu
python-dotenv
wordcloud
matplotlib
pyttsx3
